requirements.txt
mediapipe
Pillow
numpy
opencv-python
Flask
